# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Films API",
        "version": "0.1"
    },
    "paths": {
...
        "/api/critics/{filmId}": {
            "post": {
                "tags": [
                    "Critics"
                ],
                "summary": "Créer une critique pour un film",
                "operationId": "3baa3c00d1f5350fc3ade5e4ffbb1da0",
                "parameters": [
                    {
                        "name": "filmId",
                        "in": "path",
                        "description": "ID du film pour lequel la critique est créée",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Données de la critique à créer",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "score": {
                                        "description": "Score attribué à la critique",
                                        "type": "integer"
                                    },
                                    "comment": {
                                        "description": "Commentaire de la critique",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Critique créée avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "description": "ID de la critique créée",
                                            "type": "integer"
                                        },
                                        "user_id": {
                                            "description": "ID de l'utilisateur ayant créé la critique",
                                            "type": "integer"
                                        },
                                        "film_id": {
                                            "description": "ID du film concerné par la critique",
                                            "type": "integer"
                                        },
                                        "score": {
                                            "description": "Score attribué à la critique",
                                            "type": "integer"
                                        },
                                        "comment": {
                                            "description": "Commentaire de la critique",
                                            "type": "string"
                                        },
                                        "created_at": {
                                            "description": "Date et heure de création de la critique",
                                            "type": "string",
                                            "format": "date-time"
                                        },
                                        "updated_at": {
                                            "description": "Date et heure de mise à jour de la critique",
                                            "type": "string",
                                            "format": "date-time"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Accès interdit",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Vous avez déjà critiqué ce film."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "Token": []
                    }
                ]
            }
        },
        "/api/films/{id}": {
            "put": {
                "tags": [
                    "Films"
                ],
                "summary": "Mise à jour d'un film",
                "operationId": "d2fbe78b6e79c3f1cfb5067d25f9a039",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID du film à mettre à jour",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Données du film à mettre à jour",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "type": "string"
                                    },
                                    "release_year": {
                                        "type": "integer"
                                    },
                                    "length": {
                                        "type": "integer"
                                    },
                                    "description": {
                                        "type": "string"
                                    },
                                    "rating": {
                                        "type": "string"
                                    },
                                    "language_id": {
                                        "type": "integer"
                                    },
                                    "special_features": {
                                        "type": "string"
                                    },
                                    "image": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Film mis à jour avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer"
                                        },
                                        "title": {
                                            "type": "string"
                                        },
                                        "release_year": {
                                            "type": "integer"
                                        },
                                        "length": {
                                            "type": "integer"
                                        },
                                        "description": {
                                            "type": "string"
                                        },
                                        "rating": {
                                            "type": "string"
                                        },
                                        "language_id": {
                                            "type": "integer"
                                        },
                                        "special_features": {
                                            "type": "string"
                                        },
                                        "image": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Non authentifié",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Non authentifié"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Accès interdit",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Vous n'avez pas les autorisations nécessaires pour cette action."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Film non trouvé",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Le film n'existe pas."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "Token": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "Token": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}